doc: mocks & renamed
