+++ src/pages/home.tsx
@@ -4,10 +4,9 @@ import { useState } from "react";
 import { Container, Box, Button } from 'src/components/mui';
 
 import TwoLevelPieChart from 'src/components/charts/pie'
-import PieChart from 'src/components/charts/pie2'
 import PositiveAndNegativeBarChart from 'src/components/charts/bar'
 import { LineChart } from "@mui/x-charts/LineChart";
-import ModalDialog from 'src/components/tools/dialog'
+import ModalDialog from 'src/components/tools/emtionDialog'
 
 import { pagelayout, basicCenterlayout, twoColslayout, leftColumn, rightColumn } from 'src/assets/styles/commonStyles'
 
@@ -26,12 +25,12 @@ export default function HomePage() {
         </Button>
       </Box>
 
-      <ModalDialog
+      {/* <ModalDialog
         open={activeDialog === "input_feel"}
         handleClose={() => setActiveDialog(null)}
         title="Your Feel">
         <p>這是children內容。</p>
-      </ModalDialog>
+      </ModalDialog> */}
 
 
       <Box sx={{ height: "300px", width: "100%" }}>
@@ -56,7 +55,7 @@ export default function HomePage() {
           <p>is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries.</p>
         </Box>
         <Box sx={{ ...leftColumn, height: "100vh", width: "100%" }}>
-          <PieChart />
+          {/* <PieChart /> */}
         </Box>
       </Box>
