--- src/pages/test.tsx
@@ -1,12 +1,0 @@
-import * as React from 'react';
-import { Container, Box, Button } from 'src/components/mui';
-import LanguageSwitcher from 'src/components/test/languageSwitcher'
-
-export default function TestPage() {
-
-    return (
-        <Container>
-            <LanguageSwitcher />
-        </Container>
-    )
-}
